+++ CHANGES.md
@@ -1,8 +1,8 @@
 JAMES Examples Module Changes
 =============================
 
-Version 1.0 (under development)
--------------------------------
+Version 1.0 (17/06/2015)
+------------------------
 
  - Compatible with core module v1.0.
  - Extended several examples.